requirements.txt
absl-py==2.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
executing==2.2.0
fastapi==0.115.12
flatbuffers==25.2.10
fonttools==4.58.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-pasta==0.2.0
grpcio==1.72.1
h11==0.16.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
kagglehub==0.3.12
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
narwhals==1.41.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.6.11
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
